Move dialog to layout
